Move light modules to x-pack
