+++ metricbeat/beater/metricbeat.go
@@ -31,7 +31,6 @@ import (
 	"github.com/elastic/beats/libbeat/cfgfile"
 	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/libbeat/logp"
-	"github.com/elastic/beats/libbeat/paths"
 	"github.com/elastic/beats/metricbeat/mb"
 	"github.com/elastic/beats/metricbeat/mb/module"
 
@@ -70,14 +69,6 @@ func WithModuleOptions(options ...module.Option) Option {
 	}
 }
 
-// WithLightModules enables light modules support
-func WithLightModules() Option {
-	return func(*Metricbeat) {
-		path := paths.Resolve(paths.Home, "module")
-		mb.Registry.SetSecondarySource(mb.NewLightModulesSource(path))
-	}
-}
-
 // Creator returns a beat.Creator for instantiating a new instance of the
 // Metricbeat framework with the given options.
 func Creator(options ...Option) beat.Creator {
@@ -99,7 +90,6 @@ func Creator(options ...Option) beat.Creator {
 //     )
 func DefaultCreator() beat.Creator {
 	return Creator(
-		WithLightModules(),
 		WithModuleOptions(
 			module.WithMetricSetInfo(),
 			module.WithServiceName(),
